move resource server code to miscellaneous. make it more generic by introducing a config type
