+++ src/tests/end-to-end/common/test-resources.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as testServerConfig from '../setup/test-server-config';
+import { testResourceServerConfig } from 'tests/end-to-end/setup/test-resource-server-config';
 
 export function getTestResourceUrl(path: string): string {
-    return `http://localhost:${testServerConfig.port}/${path}`;
+    return `http://localhost:${testResourceServerConfig.port}/${path}`;
 }
