+++ src/tests/end-to-end/setup/test-resource-server-config.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { ResourceServerConfig } from 'tests/miscellaneous/test-resource-server/resource-server-config';
+
+export const testResourceServerConfig: ResourceServerConfig = {
+    port: 9050,
+    path: '../../end-to-end/test-resources',
+};
